enhance manual workout detail view with 1 rep max and total weight lifted, added better formatting
added support for adding a new exercise when in the middle of a workout

added support for retrieving strava detailed activity upon click in workouts list

+++ app/src/main/java/com/mcwilliams/theninjamethod/ui/workouts/manualworkoutdetail/ManualWorkoutDetailFragment.kt
@@ -2,9 +2,7 @@ package com.mcwilliams.theninjamethod.ui.workouts.manualworkoutdetail
 
 import android.os.Bundle
 import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
@@ -14,10 +12,13 @@ import com.google.android.material.textview.MaterialTextView
 import com.mcwilliams.theninjamethod.R
 import com.mcwilliams.theninjamethod.databinding.WorkoutDetailFragmentBinding
 import com.mcwilliams.theninjamethod.ui.workouts.combinedworkoutlist.model.Workout
+import com.mcwilliams.theninjamethod.ui.workouts.combinedworkoutlist.model.WorkoutSet
 import com.mcwilliams.theninjamethod.utils.extensions.fixCase
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.android.synthetic.main.workout_detail_fragment.*
+import java.text.NumberFormat
 import java.time.LocalDate
+import java.util.*
 
 @AndroidEntryPoint
 class ManualWorkoutDetailFragment : Fragment() {
@@ -25,6 +26,7 @@ class ManualWorkoutDetailFragment : Fragment() {
     private lateinit var binding: WorkoutDetailFragmentBinding
     lateinit var workout: Workout
     private val viewModel: ManualWorkoutViewModel by viewModels()
+    private var totalAmountLifted = 0
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -54,43 +56,56 @@ class ManualWorkoutDetailFragment : Fragment() {
                 "${date.dayOfWeek.name.fixCase()}, ${date.month.name.fixCase()} ${date.dayOfMonth}, ${date.year}"
 
             for (exercise in it.exercises) {
-                val exerciseRow = layoutInflater.inflate(R.layout.workout_detail_sets_row, null)
+                val exerciseRow =
+                    layoutInflater.inflate(R.layout.workout_detail_exercise_header_row, null)
                 val exerciseName =
                     exerciseRow.findViewById<MaterialTextView>(R.id.exercise_name_detail)
                 exerciseName.text = exercise.exerciseName
                 exercises_and_sets.addView(exerciseRow)
 
-                val setNumber = exerciseRow.findViewById<MaterialTextView>(R.id.set_count_detail)
-                val setRepsAndWeight =
-                    exerciseRow.findViewById<MaterialTextView>(R.id.reps_and_weight_count)
-
-                setNumber.visibility = View.GONE
-                setRepsAndWeight.visibility = View.GONE
+                totalWeightLifted(exercise.sets)
 
                 exercise.sets.forEach {
                     val setRow = layoutInflater.inflate(R.layout.workout_detail_sets_row, null)
-
-                    val exerciseNameOnSetRow =
-                        setRow.findViewById<MaterialTextView>(R.id.exercise_name_detail)
-                    exerciseNameOnSetRow.visibility = View.GONE
 
                     val setNumber = setRow.findViewById<MaterialTextView>(R.id.set_count_detail)
                     setNumber.text = it.index
                     val setRepsAndWeight =
                         setRow.findViewById<MaterialTextView>(R.id.reps_and_weight_count)
-                    setRepsAndWeight.text = "${it.weight} x ${it.reps}"
+                    setRepsAndWeight.text = "${it.weight}lbs x ${it.reps}"
+
+                    val onRepMaxTextView =
+                        setRow.findViewById<MaterialTextView>(R.id.one_rep_max_value)
+                    val oneRepMax =
+                        (it.weight.toInt() / (1.0278 - (0.0278 * it.reps.toInt()))).toInt()
+                    onRepMaxTextView.text = oneRepMax.toString() + "lbs"
 
                     exercises_and_sets.addView(setRow)
                 }
             }
+            total_weight_lifted.text =
+                NumberFormat.getNumberInstance(Locale.US).format(totalAmountLifted) + "lbs lifted"
         })
 
         viewModel.getManualWorkoutDetail(workout.id)
-
-        delete_workout.setOnClickListener {
-            viewModel.deleteWorkout()
-            Navigation.findNavController(it).popBackStack()
-        }
     }
 
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        super.onCreateOptionsMenu(menu, inflater)
+        inflater.inflate(R.menu.manual_workout_detail_menu, menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (R.id.menu_delete == item.itemId) {
+            viewModel.deleteWorkout()
+            Navigation.findNavController(binding.root).popBackStack()
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+    fun totalWeightLifted(sets: List<WorkoutSet>) {
+        sets.forEach {
+            totalAmountLifted += (it.weight.toInt() * it.reps.toInt())
+        }
+    }
 }
